Finished the profile view and did a minor refactor since the details tables for profile and earthquakedetails are the same. Now they share a table component.
