+++ src/views/Profile.js
@@ -1,15 +1,49 @@
 import React, {useState, useEffect} from 'react';
 
 // COMPONENTS //
-
+import Table from '../components/Table';
 
 const Profile = (props) => {
 
-    return (
-        <section className={'view'}>
+    const profileObject = props.data?.profile;
 
-        </section>
-    )
+    const rowMapping = [
+        {
+            label: "First name",
+            value: profileObject?.firstName
+        },
+        {
+            label: "Last name",
+            value: profileObject?.lastName
+        },
+        {
+            label: "Phone",
+            value: profileObject?.phone
+        },
+        {
+            label: "Email",
+            value: profileObject?.email
+        },
+        {
+            label: "Bio",
+            value: profileObject?.bio
+        }
+    ]
+
+    if(props.data){
+        return (
+            <section>
+                 <h2>Profile</h2>
+                 <div className={'flexRow'}>
+                     <img className={'avatar'} src={profileObject.avatarImage} alt={profileObject.firstName + ' ' + profileObject.lastName} title={profileObject.firstName + ' ' + profileObject.lastName} />
+                     <Table rowMapping={rowMapping} />
+                 </div>
+            </section>
+        )
+    } else {
+        return null;
+    }
+    
 }
 
 export default Profile;
